requirements.txt
fastapi
uvicorn[standard]
pydantic
mflux
pillow
requests
click>=8.0.0